# Conflicts:
#	src/main/java/com/tang/common/test/BBB.java

+++ src/main/java/com/tang/common/test/BBB.java
@@ -5,6 +5,6 @@ public class BBB {
 
         System.out.println(new Integer(350));
 
-        System.out.println(new Integer(800));
+        System.out.println(new Integer(1000));
     }
 }
